Turning UI into class
