+++ index.js
@@ -1,16 +1,14 @@
 'use strict'
 
+const assert = require('assert')
 const Blockly = require('blockly/blockly_compressed')
 
 const Restore = require('./libs/persist')
 const Env = require('./libs/env')
 
-// Must load this first to create Blockly.TidyBlocks
-require('./blocks/codegen')
-const {
-  createTheme,
-  createValidators
-} = require('./blocks/util')
+// Must load 'blocks/util' here so that Blockly.TidyBlocks exists before the
+// code in './blocks/*' tries to define blocks.
+const blocks = require('./blocks/blocks')
 
 // Load block files for their side effects (block definitions).
 require('./blocks/combine')
@@ -22,66 +20,73 @@ require('./blocks/transform')
 require('./blocks/value')
 
 /**
- * Blockly workspace.
+ * User interface mediator.
  */
-let TidyBlocksWorkspace = null // assigned in setup
-
-/**
- * Get the workspace or throw an Error if it has not yet been set up.
- */
-const getWorkspace = () => {
-  if (!TidyBlocksWorkspace) {
-    throw new Error('Workspace not initialized')
+class UserInterface {
+  /**
+   * Build user interface object.
+   * @param divId HTML ID of 'div' element containing workspace.
+   * @param toolboxId HTML ID of 'xml' element containing toolbox spec.
+   */
+  constructor (divId, toolboxId) {
+    blocks.createValidators()
+    const settings = this._createSettings(toolboxId)
+    this.workspace = Blockly.inject(divId, settings)
+    this.env = null
   }
-  return TidyBlocksWorkspace
-}
 
-/**
- * Get the JSON string representation of the workspace contents.
- */
-const getCode = () => {
-  const workspace = getWorkspace()
-  return Blockly.TidyBlocks.workspaceToCode(workspace)
-}
-
-/**
- * Get the object representation of the current program.
- */
-const getProgram = () => {
-  const code = getCode()
-  const json = JSON.parse(code)
-  const converter = new Restore()
-  return converter.program(json)
-}
-
-/**
- * Run the current program.
- */
-const runProgram = () => {
-  const program = getProgram()
-  const env = new Env()
-  program.run(env)
-  return env
-}
-
-/**
- * Create the JSON settings used to initialize the workspace.  Requires the DOM
- * element containing the block definitions.
- */
-const createSettings = (toolbox) => {
-  const settings = {
-    toolbox: toolbox,
-    zoom: {
-      controls: true,
-      wheel: true,
-      startScale: 1.0,
-      maxScale: 3,
-      minScale: 0.3,
-      scaleSpeed: 1.2
-    },
-    theme: createTheme()
+  /**
+   * Get the JSON string representation of the workspace contents.
+   */
+  getCode () {
+    assert(this.workspace,
+           `Workspace has not been initialized`)
+    return Blockly.TidyBlocks.workspaceToCode(this.workspace)
   }
-  return settings
+
+  /**
+   * Get the object representation of the current program.
+   */
+  getProgram () {
+    const code = this.getCode()
+    const json = JSON.parse(code)
+    const converter = new Restore()
+    return converter.program(json)
+  }
+
+  /**
+   * Run the current program, leaving state in 'this.env'.
+   */
+  runProgram () {
+    const program = this.getProgram()
+    this.env = new Env()
+    program.run(this.env)
+  }
+
+  /**
+   * Create the JSON settings used to initialize the workspace.  Requires the
+   * DOM element containing the block definitions.
+   * @param toolboxId HTML ID of 'xml' element containing toolbox spec.
+   * @returns JSON settings object.
+   */
+  _createSettings (toolboxId) {
+    toolbox = document.getElementById(toolboxId)
+    assert(toolbox,
+           `No toolbox found with ID ${toolboxId}`)
+    const theme = blocks.createTheme()
+    return {
+      toolbox,
+      theme,
+      zoom: {
+        controls: true,
+        wheel: true,
+        startScale: 1.0,
+        maxScale: 3,
+        minScale: 0.3,
+        scaleSpeed: 1.2
+      }
+    }
+  }
 }
 
 /**
@@ -89,18 +94,9 @@ const createSettings = (toolbox) => {
  * the UI and of the element that contains the block specs.
  */
 const setup = (divId, toolboxId) => {
-  createValidators()
-  const toolbox = document.getElementById(toolboxId)
-  const settings = createSettings(toolbox)
-  TidyBlocksWorkspace = Blockly.inject(divId, settings)
-  return TidyBlocksWorkspace
+  return new UserInterface(divId, toolboxId)
 }
 
 module.exports = {
-  Blockly,
-  getWorkspace,
-  getCode,
-  getProgram,
-  runProgram,
   setup
 }
